Expose tables_count and table_cell_F1 in reports
