+++ artifacts/bench-2.html
@@ -1,15 +1,15 @@
 <html><body>
 <h2>Timing</h2><table border='1'><tr><th>Mode</th><th>md_ms avg</th><th>md_ms std</th></tr>
 <tr><td>pre</td><td>0.1</td><td>0.0</td></tr>
-<tr><td>post-1S</td><td>0.6</td><td>0.1</td></tr>
-<tr><td>post-2</td><td>0.9</td><td>0.1</td></tr>
-<tr><td>python-hot</td><td>5.4</td><td>0.7</td></tr>
-<tr><td>python-cold</td><td>3118.6</td><td>82.3</td></tr>
+<tr><td>post-1S</td><td>0.7</td><td>0.1</td></tr>
+<tr><td>post-2</td><td>1.4</td><td>0.2</td></tr>
+<tr><td>python-hot</td><td>6.7</td><td>0.9</td></tr>
+<tr><td>python-cold</td><td>3434.2</td><td>130.6</td></tr>
 </table>
-<p>post-1S vs pre: 449.5%</p>
-<p>post-2 vs post-1S: 34.1%</p>
+<p>post-1S vs pre: 494.1%</p>
+<p>post-2 vs post-1S: 91.7%</p>
 <p>python-hot vs python-cold: -99.8%</p>
-<p>post-2 vs python-hot: -83.8%</p>
+<p>post-2 vs python-hot: -79.7%</p>
 <h2>Quality vs python-hot</h2><table border='1'><tr><th>Mode</th><th>CER</th><th>Token-F1</th><th>line_F1</th><th>tables_count</th><th>line_count</th><th>list_items</th><th>pipes_lines</th><th>median_pipes</th><th>max_pipes</th></tr>
 <tr><td>pre</td><td>0.001</td><td>1.000</td><td>0.986</td><td>0</td><td>36</td><td>0</td><td>16</td><td>2.5</td><td>6</td></tr>
 <tr><td>post-1S</td><td>0.001</td><td>1.000</td><td>0.986</td><td>0</td><td>36</td><td>0</td><td>16</td><td>2.5</td><td>6</td></tr>
