+++ artifacts/bench-2.json
@@ -6,23 +6,23 @@
       "timing": {
         "trials": [
           {
-            "md_ms": 0.1029
+            "md_ms": 0.138
           },
           {
-            "md_ms": 0.1223
+            "md_ms": 0.1335
           },
           {
-            "md_ms": 0.1304
+            "md_ms": 0.119
           },
           {
-            "md_ms": 0.1305
+            "md_ms": 0.123
           },
           {
-            "md_ms": 0.1044
+            "md_ms": 0.089
           }
         ],
-        "avg_ms": 0.11810000000000001,
-        "std_ms": 0.012177191794498431
+        "avg_ms": 0.12050000000000001,
+        "std_ms": 0.01718138527593163
       },
       "quality_vs_python_hot": {
         "text": {
@@ -54,23 +54,23 @@
       "timing": {
         "trials": [
           {
-            "md_ms": 0.7972
+            "md_ms": 0.8088
           },
           {
-            "md_ms": 0.5454
+            "md_ms": 0.6339
           },
           {
-            "md_ms": 0.663
+            "md_ms": 0.6921
           },
           {
-            "md_ms": 0.5558
+            "md_ms": 0.6875
           },
           {
-            "md_ms": 0.6835
+            "md_ms": 0.7571
           }
         ],
-        "avg_ms": 0.6489800000000001,
-        "std_ms": 0.09248641846238832
+        "avg_ms": 0.71588,
+        "std_ms": 0.06070395044805566
       },
       "quality_vs_python_hot": {
         "text": {
@@ -102,23 +102,23 @@
       "timing": {
         "trials": [
           {
-            "md_ms": 1.0677
+            "md_ms": 1.6325
           },
           {
-            "md_ms": 0.8722
+            "md_ms": 1.4099
           },
           {
-            "md_ms": 0.8178
+            "md_ms": 1.3968
           },
           {
-            "md_ms": 0.7978
+            "md_ms": 1.2522
           },
           {
-            "md_ms": 0.7953
+            "md_ms": 1.1705
           }
         ],
-        "avg_ms": 0.87016,
-        "std_ms": 0.10257462844193006
+        "avg_ms": 1.3723800000000002,
+        "std_ms": 0.158028648035728
       },
       "quality_vs_python_hot": {
         "text": {
@@ -150,23 +150,23 @@
       "timing": {
         "trials": [
           {
-            "md_ms": 5.2562760000114395
+            "md_ms": 8.27246300002571
           },
           {
-            "md_ms": 4.997698000011042
+            "md_ms": 6.541440999967563
           },
           {
-            "md_ms": 6.708107000008567
+            "md_ms": 5.57544500009044
           },
           {
-            "md_ms": 4.946372999995674
+            "md_ms": 6.298282999978255
           },
           {
-            "md_ms": 4.9592759999654845
+            "md_ms": 7.032727000023442
           }
         ],
-        "avg_ms": 5.373545999998441,
-        "std_ms": 0.676781122608806
+        "avg_ms": 6.744071800017082,
+        "std_ms": 0.8971930648004445
       },
       "quality_vs_python_hot": {
         "text": {
@@ -198,23 +198,23 @@
       "timing": {
         "trials": [
           {
-            "md_ms": 3161.9404
+            "md_ms": 3385.9818
           },
           {
-            "md_ms": 3056.0118
+            "md_ms": 3347.1342
           },
           {
-            "md_ms": 3039.6731
+            "md_ms": 3655.7192
           },
           {
-            "md_ms": 3260.0243
+            "md_ms": 3497.415
           },
           {
-            "md_ms": 3075.3624
+            "md_ms": 3284.7148
           }
         ],
-        "avg_ms": 3118.6024,
-        "std_ms": 82.33523091661307
+        "avg_ms": 3434.193,
+        "std_ms": 130.62784183209945
       },
       "quality_vs_python_hot": {
         "text": {
